chore: extract load-broadcast suite into src/suites/load-broadcast.ts
Part of REAL-1037.
