+++ test/e2e/realtime-check.ts
@@ -25,6 +25,7 @@ import { connection } from "./src/suites/connection.ts";
 import { loadPresence } from "./src/suites/load-presence.ts";
 import { broadcast } from "./src/suites/broadcast.ts";
 import { presence } from "./src/suites/presence.ts";
+import { loadBroadcast } from "./src/suites/load-broadcast.ts";
 
 async function runLoadPostgresChangesTests(testUser: { email: string; password: string }) {
   suite("load-postgres-changes");
@@ -134,77 +135,6 @@ async function runLoadPostgresChangesTests(testUser: { email: string; password: 
       await settle(() => latencies.length, LOAD_MESSAGES, LOAD_SETTLE_MS);
 
       return measureThroughput(latencies, LOAD_MESSAGES, "DELETE events", LOAD_DELIVERY_SLO);
-    } finally {
-      await stopClient(supabase);
-    }
-  });
-}
-
-async function runLoadBroadcastTests() {
-  suite("load-broadcast");
-
-  await sleep(RATE_LIMIT_PAUSE_MS);
-  await test("broadcast self throughput", async () => {
-    const supabase = createClient(PROJECT_URL, ANON_KEY, { realtime: REALTIME_OPTS });
-    try {
-      const event = "load";
-      const topic = randomTopic();
-      const sendTimes = new Map<number, number>();
-      const latencies: number[] = [];
-
-      const channel = supabase
-        .channel(topic, BROADCAST_CONFIG)
-        .on("broadcast", { event }, ({ payload }) => {
-          const t = sendTimes.get(payload.seq);
-          if (t !== undefined) latencies.push(performance.now() - t);
-        });
-
-      await openChannel(channel);
-
-      for (let i = 0; i < LOAD_MESSAGES; i++) {
-        sendTimes.set(i, performance.now());
-        await channel.send({ type: "broadcast", event, payload: { seq: i } });
-      }
-
-      await settle(() => latencies.length, LOAD_MESSAGES, LOAD_SETTLE_MS);
-
-      return measureThroughput(latencies, LOAD_MESSAGES, "broadcast events", LOAD_DELIVERY_SLO);
-    } finally {
-      await stopClient(supabase);
-    }
-  });
-
-  await sleep(RATE_LIMIT_PAUSE_MS);
-  await test("broadcast API endpoint throughput", async () => {
-    const supabase = createClient(PROJECT_URL, ANON_KEY, { realtime: REALTIME_OPTS });
-    try {
-      const event = "load";
-      const topic = randomTopic();
-      const sendTimes = new Map<number, number>();
-      const latencies: number[] = [];
-
-      const channel = supabase
-        .channel(topic, BROADCAST_CONFIG)
-        .on("broadcast", { event }, ({ payload }) => {
-          const t = sendTimes.get(payload.seq);
-          if (t !== undefined) latencies.push(performance.now() - t);
-        });
-
-      await openChannel(channel);
-
-      await Promise.all(Array.from({ length: LOAD_MESSAGES }, async (_, i) => {
-        sendTimes.set(i, performance.now());
-        const res = await fetch(`${PROJECT_URL}/realtime/v1/api/broadcast`, {
-          method: "POST",
-          headers: BROADCAST_API_HEADERS,
-          body: JSON.stringify({ messages: [{ topic, event, payload: { seq: i } }] }),
-        });
-        if (!res.ok) throw new Error(`Broadcast API returned ${res.status}`);
-      }));
-
-      await settle(() => latencies.length, LOAD_MESSAGES, LOAD_SETTLE_MS);
-
-      return measureThroughput(latencies, LOAD_MESSAGES, "broadcast API events", LOAD_DELIVERY_SLO);
     } finally {
       await stopClient(supabase);
     }
@@ -1134,7 +1064,7 @@ const descriptors: SuiteDescriptor[] = [
   connection,
   { name: "load-postgres-changes", label: "load-postgres-changes", needsDb: true, run: ({ testUser }) => runLoadPostgresChangesTests(testUser) },
   loadPresence,
-  { name: "load-broadcast", label: "load-broadcast", needsDb: false, run: () => runLoadBroadcastTests() },
+  loadBroadcast,
   loadBroadcastFromDb,
   loadBroadcastReplay,
   broadcast,
